- fix all currently reported UI use in Model - add CommandArgs.cs and CommandLine.cs as files not allowed to use UI classes - add a TODO to extend this static analysis to catch more in the future, after this PR is merged
