+++ pwiz_tools/Skyline/Controls/ImmediateWindow.cs
@@ -27,6 +27,7 @@ using System.Windows.Forms;
 using pwiz.Skyline.Alerts;
 using pwiz.Skyline.Model.Tools;
 using pwiz.Skyline.Properties;
+using pwiz.Skyline.ToolsUI;
 using pwiz.Skyline.Util;
 
 namespace pwiz.Skyline.Controls
